+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Fixed
+
+* Fixed the submission of application-type jobs. (#31, #32, #33)
+
 ## Version v0.1.4 - 2023-08-21
 
 ### Fixed